Fix bad reference to 'buster' global
Running 'buster static' was failing with an error such as:

  Error loading configuration /path/to/buster.js
  buster is not defined

When using buster-core within a browser context, everything is available via a
'buster' global variable. However this variable is not available in a node
context, which is where buster-configuration is used. buster-core is required
at the top of group.js as 'B', so we must use this instead of 'buster'.